+++ relearn/source/structure/Partition.cpp
@@ -26,7 +26,7 @@ Partition::Partition(const size_t num_ranks, const size_t my_rank)
      * We choose power of 8 as every domain subdivision creates 8 local_subdomains (in 3d).
      */
     const double smallest_exponent = std::ceil(std::log(num_ranks) / std::log(8.0));
-    level_of_subdomain_trees = static_cast<size_t>(smallest_exponent);
+    level_of_subdomain_trees = static_cast<std::uint16_t>(smallest_exponent);
     total_number_subdomains = 1ULL << (3 * level_of_subdomain_trees); // 8^level_of_subdomain_trees
 
     // Every rank should get at least one subdomain
